tratamentos de objetos opcionais
