+++ OramaCadastroSwift/Classes/OpenAPIs/Models/ContaBancaria.swift
@@ -9,31 +9,31 @@ import Foundation
 
 /** Modelo de Conta bancária no sistema bancário brasileiro. Deve ser enviado somente uma conta bancária na lista de &#39;contaBancaria&#39;. */
 
-@objc public class ContaBancaria: NSObject, Codable { 
-
+@objc public class ContaBancaria: NSObject, Codable {
+    
     public enum Tipo: String, Codable {
         case cc = "CC"
         case cp = "CP"
     }
     /** Banco da conta, string com o número do banco */
-    public var banco: String
+    public var banco: String?
     /** Tipo da conta bancária. O tipo deve ser conta corrente (CC) ou conta poupança (CP). */
-    public var tipo: Tipo = .cc
+    public var tipo: Tipo? = .cc
     /** Agência da conta bancária */
-    public var agencia: String
+    public var agencia: String?
     /** Número da conta bancária sem o dígito verificador */
-    public var conta: String
+    public var conta: String?
     /** Digito verificador da conta bancária */
-    public var digito: String
+    public var digito: String?
     /** Informação que define se é uma conta conjunta. Caso seja, deve ser definido como true. */
-    public var contaConjunta: Bool = false
+    public var contaConjunta: Bool? = false
     public var contaConjuntaNum: NSNumber? {
         get {
             return contaConjunta as NSNumber?
         }
     }
     /** Informação se o segundo participante (co-titular) é o titular da conta, caso não seja, o primeiro participante quem esta preenchendo a conta é o titular. */
-    public var segundoParticipanteTitular: Bool = false
+    public var segundoParticipanteTitular: Bool? = false
     public var segundoParticipanteTitularNum: NSNumber? {
         get {
             return segundoParticipanteTitular as NSNumber?
@@ -41,8 +41,8 @@ import Foundation
     }
     /** CPF do co-titular caso seja conta conjunta. CPF deve ser válido. O CPF tem que ter os 11 dígitos com a máscara incluindo os pontos e hífen.  - Obrigatório caso a conta bancária seja conta conjunta */
     public var cpfCotitular: String?
-
-    public init(banco: String, tipo: Tipo, agencia: String, conta: String, digito: String, contaConjunta: Bool, segundoParticipanteTitular: Bool, cpfCotitular: String?) {
+    
+    public init(banco: String?, tipo: Tipo?, agencia: String?, conta: String?, digito: String?, contaConjunta: Bool?, segundoParticipanteTitular: Bool?, cpfCotitular: String?) {
         self.banco = banco
         self.tipo = tipo
         self.agencia = agencia
@@ -52,5 +52,5 @@ import Foundation
         self.segundoParticipanteTitular = segundoParticipanteTitular
         self.cpfCotitular = cpfCotitular
     }
-
+    
 }
