+++ OramaCadastroSwift/Classes/OpenAPIs/Models/DadosPatrimonial.swift
@@ -10,7 +10,7 @@ import Foundation
 /** Informação sobre o Patrimônio do usuário */
 
 @objc public class DadosPatrimonial: NSObject, Codable {
-
+    
     public enum FaixaAplicacoes: String, Codable {
         case _0 = "0"
         case _1 = "1"
@@ -35,14 +35,14 @@ import Foundation
     public var valorAplicacoes: Double?
     /** Código do valor que representa a renta bruta mensal. Cada número representa um intervalo de valor em reais (R$). Por exemplo, &#39;0&#39;: &#39;Nenhum&#39;, &#39;1&#39;: &#39;Até R$5.000,00&#39;, &#39;2&#39;: &#39;De R$5.000,01 à R$10.000,00&#39;, &#39;3&#39;: &#39;De R$10.000,01 à R$20.000,00&#39;, &#39;4&#39;: &#39;De R$20.000,01 à R$30.000,00&#39;, &#39;5&#39;: &#39;De R$30.000,01 à R$100.000,00&#39;, &#39;6&#39;: &#39;Acima de R$100.000,00&#39; */
     public var faixaRendimentos: FaixaRendimentos?
-    /** Valor total em R$ incluindo móveis, imóveis e outros bens do usuário  - Orbigatório caso &#39;faixaAplicacoes&#39; e &#39;faixaRendimentos&#39; seja &#39;0 - Nenhum&#39; */
+    /** Valor total em R$ incluindo móveis, imóveis e outros bens do usuário  - Obrigatório caso &#39;faixaAplicacoes&#39; e &#39;faixaRendimentos&#39; seja &#39;0 - Nenhum&#39; */
     public var valorOutrosBens: Double?
-
+    
     public init(faixaAplicacoes: FaixaAplicacoes?, valorAplicacoes: Double?, faixaRendimentos: FaixaRendimentos?, valorOutrosBens: Double?) {
         self.faixaAplicacoes = faixaAplicacoes
         self.valorAplicacoes = valorAplicacoes
         self.faixaRendimentos = faixaRendimentos
         self.valorOutrosBens = valorOutrosBens
     }
-
+    
 }
