+++ OramaCadastroSwift/Classes/OpenAPIs/Models/Endereco.swift
@@ -9,8 +9,8 @@ import Foundation
 
 /** Definição de endereço */
 
-@objc public class Endereco: NSObject, Codable { 
-
+@objc public class Endereco: NSObject, Codable {
+    
     public enum Uf: String, Codable {
         case ac = "AC"
         case al = "AL"
@@ -41,21 +41,21 @@ import Foundation
         case to = "TO"
     }
     /** Código Postal Brasileiro */
-    public var cep: String
+    public var cep: String?
     /** logradouro do endereço */
-    public var logradouro: String
+    public var logradouro: String?
     /** bairro do endereço, preferencialmente lexicográficamente igual a descrição obtida pelo uso do CEP */
-    public var bairro: String
+    public var bairro: String?
     /** Unidade da Federação do endereço */
-    public var uf: Uf
+    public var uf: Uf?
     /** Município do endereço. Formato é o nome lexicograficamente igual a descrição do IBGE ou o código de cidade completo do IBGE */
-    public var cidade: String
+    public var cidade: String?
     /** Número do logradouro no endereço */
-    public var numero: String
+    public var numero: String?
     /** Complemento do Endereço */
     public var complemento: String?
-
-    public init(cep: String, logradouro: String, bairro: String, uf: Uf, cidade: String, numero: String, complemento: String?) {
+    
+    public init(cep: String?, logradouro: String?, bairro: String?, uf: Uf?, cidade: String?, numero: String?, complemento: String?) {
         self.cep = cep
         self.logradouro = logradouro
         self.bairro = bairro
@@ -64,5 +64,5 @@ import Foundation
         self.numero = numero
         self.complemento = complemento
     }
-
+    
 }
