--- OramaCadastroSwift/Classes/OpenAPIs/Models/RetornoLoginObjeto.swift
@@ -1,32 +1,0 @@
-//
-// RetornoLoginObjeto.swift
-//
-// Generated by openapi-generator
-// https://openapi-generator.tech
-//
-
-import Foundation
-
-
-
-@objc public class RetornoLoginObjeto: NSObject, Codable { 
-
-    /** Campo texto de preenchimento obrigatório para o preenchimento do nome completo do cliente. */
-    public var nome: String?
-    /** Telefone celular contendo o código de área e os nove digitos. */
-    public var celular: String?
-    /** CPF deve ser válido e não ter sido usado anteriormente. O CPF tem que ter os 11 dígitos com a máscara incluindo os pontos e hífen. */
-    public var cpf: String?
-    public var email: String?
-    /** Data de nascimento da pessoa no formato YYYY-MM-DD (cf. RFC 3339, section 5.8) */
-    public var dataNascimento: Date?
-
-    public init(nome: String?, celular: String?, cpf: String?, email: String?, dataNascimento: Date?) {
-        self.nome = nome
-        self.celular = celular
-        self.cpf = cpf
-        self.email = email
-        self.dataNascimento = dataNascimento
-    }
-
-}
