--- OramaCadastroSwift/Classes/OpenAPIs/Models/RetornoNovaSenha.swift
@@ -1,22 +1,0 @@
-//
-// RetornoNovaSenha.swift
-//
-// Generated by openapi-generator
-// https://openapi-generator.tech
-//
-
-import Foundation
-
-
-
-@objc public class RetornoNovaSenha: NSObject, Codable { 
-
-    public var cpf: String?
-    public var email: String?
-
-    public init(cpf: String?, email: String?) {
-        self.cpf = cpf
-        self.email = email
-    }
-
-}
